Revert public API changes
