+++ dokka-subprojects/plugin-base/src/test/kotlin/content/AllTypesPageTest.kt
@@ -5,8 +5,8 @@
 package content
 
 import matchers.content.*
+import org.jetbrains.dokka.base.pages.AllTypesPageNode
 import org.jetbrains.dokka.base.testApi.testRunner.BaseAbstractTest
-import org.jetbrains.dokka.pages.AllTypesPageNode
 import org.jetbrains.dokka.pages.RootPageNode
 import utils.withAllTypesPage
 import kotlin.test.Test
